# docs/openapi/CampusHelpATW.openapi.yaml
openapi: 3.0.3
info:
  title: CampusHelpATW API
  version: 0.1.0
  description: |
    CampusHelpATW 微信小程序后端接口（FastAPI 实现基线）。
    实时聊天 WebSocket 端点：
    - `/api/v1/ws/tasks/{task_id}?token={jwt}`
    全局未读通知 WebSocket 端点：
    - `/api/v1/ws/me/notifications?token={jwt}`
    全局通道事件：
    - `unread_snapshot` / `chat_unread`（聊天未读）
    - `task_event`（任务关键节点）
servers:
  - url: http://127.0.0.1:3000
    description: Local
  - url: https://api.example.com
    description: Production (placeholder)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: 任务列表
      description: 返回任务列表，并附带 `publisher_user` / `acceptor_user` 简要信息。
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
        - in: query
          name: include_unread
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 列表结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [tasks]
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/posts:
    get:
      tags: [posts]
      summary: 帖子列表
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 帖子列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [posts]
      summary: 发帖
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '200':
          description: 发帖结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/reports:
    post:
      tags: [reports]
      summary: 提交举报
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: 举报结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/v1/admin/reports:
    get:
      tags: [admin]
      summary: 管理端举报列表
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 举报列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/reports/{report_id}/handle:
    post:
      tags: [admin]
      summary: 管理端处理举报
      parameters:
        - in: path
          name: report_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportHandleRequest'
      responses:
        '200':
          description: 处理结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/tasks/{task_id}/arbitrate:
    post:
      tags: [admin]
      summary: 管理端任务仲裁
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminArbitrateRequest'
      responses:
        '200':
          description: 仲裁结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/task-categories:
    get:
      tags: [admin]
      summary: 管理端任务分类列表
      parameters:
        - in: query
          name: is_active
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 分类列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [admin]
      summary: 管理端创建任务分类
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCategoryCreateRequest'
      responses:
        '200':
          description: 创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/task-categories/{category_id}:
    patch:
      tags: [admin]
      summary: 管理端更新任务分类
      parameters:
        - in: path
          name: category_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCategoryPatchRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/system-configs:
    get:
      tags: [admin]
      summary: 管理端系统配置列表
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/admin/system-configs/{config_key}:
    put:
      tags: [admin]
      summary: 管理端新增或更新系统配置
      parameters:
        - in: path
          name: config_key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfigUpsertRequest'
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
      required: [code, message]
...
    TaskCreateRequest:
      type: object
      required: [title, description, category, deadline_at]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location_text:
          type: string
        reward_amount:
          type: number
        reward_type:
          type: string
        deadline_at:
          type: string
          format: date-time
...
    PostCreateRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
...
    ReportCreateRequest:
      type: object
      required: [target_type, target_id, reason_code]
      properties:
        target_type:
          type: string
        target_id:
          type: integer
        reason_code:
          type: string
        reason_text:
          type: string
    ReportHandleRequest:
      type: object
      required: [action, result]
      properties:
        action:
          type: string
        result:
          type: string
    TaskCategoryCreateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        sort_order:
          type: integer
          default: 0
        is_active:
          type: boolean
          default: true
    TaskCategoryPatchRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        sort_order:
          type: integer
        is_active:
          type: boolean
    SystemConfigUpsertRequest:
      type: object
      required: [config_value]
      properties:
        config_value:
          description: 任意 JSON 值
        description:
          type: string
    AdminArbitrateRequest:
      type: object
      required: [decision, reason]
      properties:
        decision:
          type: string
        reason:
          type: string